rename all_songs to show_songs
