+++ App/templates/add_song.html
@@ -24,7 +24,7 @@
       <button class="btn btn-primary" type="submit">Create</button>
       
       <!-- This link in the form of a button allows the user to go back to the list of songs -->
-      <a class="btn btn-secondary" href="{% url 'App:all_songs' %}">Cancel</a>
+      <a class="btn btn-secondary" href="{% url 'App:show_songs' %}">Cancel</a>
       
     </form>
   </div>
