+++ App/templates/base.html
@@ -72,7 +72,7 @@
             <!-- only show links to songs and playlists once the user is authenticated -->
             {% if user.is_authenticated %}
               <li class="nav-item">
-                <a class="nav-link" href="{% url 'App:all_songs' %}">Songs</a>
+                <a class="nav-link" href="{% url 'App:show_songs' %}">Songs</a>
               </li>
               <li class="nav-item mx-2 dropdown">
                 <a class="nav-link dropdown-toggle" href="#" id="navbarPlaylistDropdown" role="button" data-bs-toggle="dropdown">Playlists</a>
