+++ App/templates/play_spotify_song.html
@@ -31,7 +31,7 @@
   <!-- this row covers all 12 columns and links back to the list of all songs -->
   <div class="row">
     <div class="col-12">
-      <p class="text-center mt-4"><a href="{% url 'App:all_songs' %}">Back to Song List</a></p>
+      <p class="text-center mt-4"><a href="{% url 'App:show_songs' %}">Back to Song List</a></p>
     </div>
   </div>
