+++ App/templates/show_songs.html
@@ -4,7 +4,7 @@
 
 {% block extra_css %}
   <!-- special styling needed for this page -->
-  <link rel="stylesheet" href="{% static 'css/all_songs.css' %}">
+  <link rel="stylesheet" href="{% static 'css/show_songs.css' %}">
 {% endblock %}
 
 {% block content %}
@@ -109,5 +109,5 @@
 
 <!-- need javascript to submit the form when any info items have changed -->
 {% block extra_js %}
-    <script src={% static 'js/all_songs.js' %}></script>
+    <script src={% static 'js/show_songs.js' %}></script>
 {% endblock %}
