+++ App/urls.py
@@ -8,7 +8,8 @@ app_name = "App"
 urlpatterns = [
     path("", views.home, name="home"),
     path("home/", views.home, name="home"),   
-    path("songs/", views.all_songs, name="all_songs"),
+    path("songs/", views.show_songs, name="show_songs"),
+    path("songs/<int:song_id>", views.show_songs, name="show_songs"),
     path("addSong/", views.add_song, name="add_song"),
     path("accounts/profile/", views.user_profile, name='user_profile'),
     path('playSong/<int:song_id>', views.play_song, name="play_song"),
