+++ App/views/__init__.py
@@ -1,10 +1,10 @@
 # import the files with all the views so django can find them 
 from .home import *
-from .all_songs import *
 from .all_playlists import *
 from .play_song import *
 from .play_song_list import *
 from .playlist_crud import *
+from .show_songs import *
 from .song_crud import *
 from .spotify_views import *
 from .user_access import *
